Symlink script: allow skip confirm for CI

+++ scripts/setup_renode.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+# # Parameters
+#
+# * `-y` - skip confirmation prompt (e.g., for CI)
+
 # Detect renode location
 RENODE="${RENODE:-$(which renode)}"
 RENODE_DIR="$(dirname $(which renode))"
@@ -13,7 +17,9 @@ VP_PYTHON_PERIPHERALS="$BASEDIR/../vp/devel/python_peripherals"
 # Check if symlink exists
 if [ ! -h "$REQUIRED_SYMLINK" ]; then
     echo "Symlink not found"
-    read -p "Create symlinks and continue? (y/N): " confirm && [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]] || exit 1
+    if [[ ! $* == *-y* ]]; then
+        read -p "Create symlinks and continue? (y/N): " confirm && [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]] || exit 1
+    fi
     # Symlink the updated Python peripheral to Renode search directory
     ln -s $(readlink -f "$VP_PYTHON_PERIPHERALS/DLA.py") "$REQUIRED_SYMLINK"
     echo "Created Symlink at \"$REQUIRED_SYMLINK\""
